presidio-image-redactor/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[project]
name = "presidio-image-redactor"
version = "0.0.57"
description = "Presidio image redactor package"
authors = [{name = "Presidio", email = "presidio@microsoft.com"}]
license = "MIT"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
keywords = ["presidio_image_redactor"]
urls = {Homepage = "https://github.com/Microsoft/presidio"}
readme = "README.md"
requires-python = ">=3.10,<3.14"

dependencies = [
    "pillow (>=9.0)",
    "pytesseract (>=0.3.7,<0.4)",
    "presidio-analyzer (>=1.9.0)",
    "matplotlib (>=3.6)",
    "pydicom (>=2.3.0)",
    "pypng (>=0.20220715.0)",
    "azure-ai-formrecognizer (>=3.3.0,<4.0.0)",
    "opencv-python (>=4.0.0,<5.0.0)",
    "python-gdcm (>=3.0.24.1)",
    "spacy >=3.8.7,<4.0.0 ; python_version >= '3.13'"]

[project.optional-dependencies]
server = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
]

[tool.poetry.group.dev.dependencies]
pip = "*"
ruff = "*"
pytest = "*"
pytest-cov = "*"
pytest-mock = "*"
python-dotenv = "*"
pre_commit = "*"
diff-cover = "*"

[tool.coverage.run]
relative_files = true
source = ["."]
